Initial scraping file
This script file contains an early version of the snscrape algorithm to scrape the given Twitter data. Some examples of EDA and data cleaning are also shown, but this will be updated with more depth at a later point. The script needs to be more organised into functions as opposed to hard-coded lines, and further analysis and visualisation is to be conducted.